pyproject.toml
[project]
name = "pidroid"
version = "5.26.1"
description = "Pidroid Discord bot for TheoTown"
authors = [{ name = "JustAnyone", email = "dominykas@svetikas.lt" }]
requires-python = "~=3.12,<3.14"
readme = "README.md"
license = "MIT"
dependencies = [
    "discord-py~=2.6.4",
    "python-dateutil>=2.9.0.post0,<3",
    "sqlalchemy>=2.0.43,<3",
    "asyncpg>=0.30.0,<0.31",
    "jishaku>=2.5.2,<3",
    "aiohttp>=3.13.2,<4",
    "pytz~=2025.2",
    "psutil>=7.0.0,<8",
    "pillow>=11.3.0,<12",
    "aiofiles>=24.1.0,<25",
    "aiodbm>=0.4.2,<0.5",
    "aiocron~=2.1",
    "emoji>=2.12.1,<3",
    "alembic>=1.16.5,<2",
    "python-dotenv>=1.0.1,<2",
    "dill>=0.4.0,<0.5",
    "pynacl>=1.6.0,<2",
    "h2>=4.2.0,<5",
    "greenlet>=3.2.4,<3.3",
    "pyjwt>=2.10.1",
    "cryptography>=45.0.7,<45.1",
    "faststream[rabbit]>=0.6.4",
]

[project.optional-dependencies]
uvloop = ["uvloop>=0.19.0,<0.20"]

[project.scripts]
pidroid = "pidroid.main:main"
main = "pidroid.main:main"
start = "pidroid.main:main"
migrate = "pidroid.main:migrate"


[dependency-groups]
dev = [
    "coverage>=7.10.0,<7.11",
    "pytest>=8.4.2,<9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.basedpyright]
ignore = ["alembic/versions"]
reportMissingTypeStubs = false
reportUnusedCallResult = false
pythonVersion = "3.12"
